COZMO-11225 cozmo2 no unity (#4574)
* Only deploy cozmoRobot2 if device api > 24

We now exit with a helpful error if an attempt is made to deploy the robot process
to a device that is not running at least Android 7 (sdk version 24).

* Remove Unity build dependency

- Fix `stage-assets.sh` script to use `copyResources.py` to copy assets from
  source locations

- Minor modifications to allow skipping sound assets (not currently needed for v2)

- Generate `allAssetsHash.txt` file using algorithm from csharp implemented in bash

- Skip unity build step when building cozmo v2 target from configure.py

- Add `deploy-assets.sh` script to push assets to device

  The app will search for assets in `/sdcard/Android/data/com.anki.cozmoengine/files/assets/`

- Call `stage-assets.sh` or `deploy-assets.sh` from `configure.py`, depending
  on build commands

These changes are not an ideal asset-build workflow, but are sufficient to remove
Unity as a build requirement for running Cozmov2 (Standalone). Note that deleting the
app from the phone will remove the assets. The `run-engine` script will re-deploy if necessary.
Also, if assets change in the repo and a new version of cozmoEngine standalone is installed,
the new assets may not be deployed.  In this case, call `standalone-apk/stage-assets.sh -f`
to force the new assets to be staged.

* Add script to add device IP to cozmo2Viz webots world

This automates the step of determining the IP address of a connected
android device and modifying the cozmo2Viz.wbt file to include that
info for remote visualization.

+++ project/build-scripts/copyResources.py
@@ -63,7 +63,7 @@ class CopyResources(object):
                         action='store', default=None, help='where engine resources are located')
     parser.add_argument('--unityAssetsPath', dest='unityAssetsPath', required=True,
                         action='store', default=None, help='where unity assets are located')
-    parser.add_argument('--soundBanksPath', dest='soundBanksPath', required=True,
+    parser.add_argument('--soundBanksPath', dest='soundBanksPath', required=False,
                         action='store', default=None, help='where sound banks are located')
     parser.add_argument('--copyMethod', dest='copyMethod', action='store', default='rsync',
                         choices=('rsync', 'copy'), required=True,
@@ -190,21 +190,22 @@ class CopyResources(object):
       return False
 
     # sound banks
-    soundBanksPlatform = ''
-    if self.options.platform == 'ios':
-      soundBanksPlatform = 'iOS'
-    elif self.options.platform == 'mac':
-      soundBanksPlatform = 'Mac'
-    elif self.options.platform == 'android':
-      soundBanksPlatform = 'Android'
+    if self.options.soundBanksPath:
+      soundBanksPlatform = ''
+      if self.options.platform == 'ios':
+        soundBanksPlatform = 'iOS'
+      elif self.options.platform == 'mac':
+        soundBanksPlatform = 'Mac'
+      elif self.options.platform == 'android':
+        soundBanksPlatform = 'Android'
 
-    soundBanksPath = os.path.join(cozmoResourcesPath, 'sound')
-    if os.path.isdir(soundBanksPath):
-      ankibuild.util.File.rm_rf(soundBanksPath)
-    if not ankibuild.util.File.cptree(os.path.join(self.options.soundBanksPath, soundBanksPlatform), soundBanksPath):
-      self.log.error("error copying {0} to {1}".format (
-        os.path.join(self.options.soundBanksPath, soundBanksPlatform), soundBanksPath))
-      return False
+      soundBanksPath = os.path.join(cozmoResourcesPath, 'sound')
+      if os.path.isdir(soundBanksPath):
+        ankibuild.util.File.rm_rf(soundBanksPath)
+      if not ankibuild.util.File.cptree(os.path.join(self.options.soundBanksPath, soundBanksPlatform), soundBanksPath):
+        self.log.error("error copying {0} to {1}".format (
+          os.path.join(self.options.soundBanksPath, soundBanksPlatform), soundBanksPath))
+        return False
 
     return True
 
@@ -262,18 +263,19 @@ class CopyResources(object):
     ankibuild.util.File.execute(args)
 
     # sound banks
-    soundBanksPlatform = ''
-    if self.options.platform == 'ios':
-      soundBanksPlatform = 'iOS/'
-    elif self.options.platform == 'mac':
-      soundBanksPlatform = 'Mac/'
-    elif self.options.platform == 'android':
-      soundBanksPlatform = 'Android/'
-
-    soundBanksPath = os.path.join(cozmoResourcesPath, 'sound')
-    args = baseargs[:]
-    args += [os.path.join(self.options.soundBanksPath, soundBanksPlatform), soundBanksPath]
-    ankibuild.util.File.execute(args)
+    if self.options.soundBanksPath:
+      soundBanksPlatform = ''
+      if self.options.platform == 'ios':
+        soundBanksPlatform = 'iOS/'
+      elif self.options.platform == 'mac':
+        soundBanksPlatform = 'Mac/'
+      elif self.options.platform == 'android':
+        soundBanksPlatform = 'Android/'
+  
+      soundBanksPath = os.path.join(cozmoResourcesPath, 'sound')
+      args = baseargs[:]
+      args += [os.path.join(self.options.soundBanksPath, soundBanksPlatform), soundBanksPath]
+      ankibuild.util.File.execute(args)
 
     return True
